requirements.txt
praw
anthropic
openai
pyyaml
beautifulsoup4
lxml
python-dotenv
prawcore
Image
requests